@@ -1,16 +1,3 @@
-module.exports = {
-    AUTOTRANSFER_INTERVAL_SECONDS: 1,
-    AUTOTRANSFER_FROM: 'auto',
-    AUTOTRANSFER_TO:
-        'BHGaoDov6gsuHbfk2Tc0cAyHABw3hoKS2Cv1uBpA+/nVc1JikV6IxqEZ/5NlizPGFpvMtONMyBeJcXOIb4Jdnjk=',
-
-    // GUI_URL: 'http://localhost:8080/gui/',
-    // GUI_USERNAME: 'username',
-    // GUI_PASSWORD: 'password',
-
-    // AUTOREMOVE_INTERVAL_SECONDS: 15,
-    // AUTOREMOVE_TORRENTS_MAX_AMOUNT: 15,
-    // AUTOREMOVE_TORRENT_MAX_SIZE_GB: 25,
-    // PEERS_FILTER_INTERVAL_SECONDS: 5,
-    // AUTOCONFIG_ENABLE: true,
+export default {
+    AUTOTRANSFER_ENABLED: true
 }
